Implementation of the Pico IR Firmware and API server
Communicates with Pico IR over serial port (/dev/serial/by-id/...).
Opening it retried on a best effort basis, if the device is disconnected
temporarily, the server should survive.

A Power-on hack is implemented to get the speakers into an "On" state,
at the cost of a reset if they were on already. Don't have a better way
because it's only a toggle command.

Otherwise audio inputs can be set or raw commands be sent (u8).